Bump Example Product to 1.8.7-alpha
This is because we did not release example product for some time due to no changes at all happening in it.
However we have decided that it is best to keep bumping example product to keep in sync with OpsManager

The reason we decided it is okay to have no-op bumps is that the example product is used for documentation
and other teams look at it when they want to know what a 1.8.X compatible product looks like.
We also discussed the fact that bumping it does not have any consequences since no one deploys it besides us.

[#132006407]