Show all accessors for generated cert in webserver
In newer versions of the example product, we display all the values for
all the accessors for generated_rsa_cert_credentials. For the
certificate rotation system spec we depend on these values being present
in the previous and current example product. Since we're backporting
this feature and spec to 2.1, we need the 2.0 example product to render
these values.

[#158471123] Operator should be able to rotate to new CA for BOSH DNS Healthiness certs that is valid for 4 years