Add job to test BOSH DNS healthiness during upgrade
A fairly simple "source" and "destination" job that can be linked
together with BOSH DNS provides the ability to detect disfunctional
BOSH DNS deployments, particularly around healthiness during upgrades.

The source job is attached to web_server, and the destinations are the no_op.
Initial deploys skip the tests, since no_op isn't deployed until after web_server,
but on subsequent deploys the testing system
deployed on the source validates that the destination nodes still appear healthy.
The advantage of pinging a "later deployed" node is that you can guarentee that
an older version of DNS can interoperate with a newer version of DNS, especially
around the healthiness features.