Changes HTTP verifier URL to something we control
`https://www.vmware.com` has started returning 403's. Why? Who knows!
But it has been causing all of our specs in CI to fail.

Turns out that UAA has a perfectly servicable web server that will
return 200s for requests, so we're using that in our verifier.

Because other parts of the system are making calls out to off-site
IAASs, we aren't losing the network-connectivity verification that this
check _was_ providing. Would be good to be able to keep this extra
check, but -because we run in Nimbus-, our options are really limited.

Signed-off-by: Brian Upton <bupton@vmware.com>