Use UAA health check, rather than login
vmware.com is now returning 403's. This makes it unsuitable for our HTTP
verifier.

The UAA health check endpoint _should_ always return a 200, assuming that
nginx and UAA are up. If neither are up, then you're not interacting
with the Ops Manager UI, and aren't going to get to run this check
anyway.

Signed-off-by: Brian Upton <bupton@vmware.com>