Use UAA health check, rather than login
On vSphere deployments, we enable SAML. This makes the UAA login
endpoint return a 302.

However, the UAA health check endpoint _should_ always return a 200,
assuming that nginx and UAA are up. If neither are up, then you're
not interacting with the Ops Manager UI, and aren't going to get to
run this check anyway.

Signed-off-by: Brian Upton <bupton@vmware.com>