Modifies example-runtime-product as value testbed
This commit adds new configuration for the Example Runtime Config tile
that serves to exercise the validation of the in-the-database Tile
Properties.

Build like so, and upload to the Ops Manager:

```
~/workspace/ops-manager-example $ ./build-tile --type=runtime-config-only
```

Currently known issues:
* First-run validation doesn't work... it always thinks that something
  is invalid
* Data in the DB isn't blown away after the tile is removed.
* Something weird is happening with manifest generation. Something like
  changes to the defaults aren't making it out to the manifest? Unclear,
  but unrelated to getting the basic property validation work done.

After the Story is done, feel free to delete this branch and destroy all
this work.

[#186997240] 🚀 Convert installation.yml => DB