Removes secret props to have zero config required
This commit removes the secret properties so that I don't have to
manually configure any properties in the tile. The secret properties
work (though there are some unanswered questions about their
implementation details), so we're not removing them because they don't
work... just because it's nice to be able to easily spot when properties
that _should_ be configured on tile import are not.